requirements.txt
zope.interface
python-openstackclient
python-keystoneclient
python-novaclient
python-neutronclient
python-cinderclient
testtools
fixtures
coverage
